Disable spellcheck in editor
**what is the change?:**
Add `spellCheck="false"` to the `contenteditable` container of the
editor.

**why make this change?:**
In Firefox and probably other places, words that are flagged by the
browser as misspelled will have a dotted underline. This really stands
out in the editor and is distracting, not relevant for code.

**test plan:**
Manual testing - Flarnie will insert screenshots into the PR.

**issue:**
https://github.com/FormidableLabs/react-live/issues/30