Add tabMode prop
This commit adds a new prop for LiveEditor - - which lets you define how the tab key should work.

If you pass the value indentation (the default), everything will work as it has. If you pass focus, however, you'll be able to tab on to the next or previous item in the tab order